Add files via upload
In this assignment I will explore two implementations of the ADT list.  The first implementation will use an array and the second will use a linked structure. In both cases I will create new methods that work directly with the implementations. I will implement a reverse method that will reverse the order of the items in the list.  I will also implement a cycle method that will move the first item in the list to the last position.